% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BackSelection.R
\name{BackSelection}
\alias{BackSelection}
\title{Build the model through backward selection}
\usage{
BackSelection(data, Type, variables, continuous, TD, NL, m = 1, p = 2,
  alpha_back = 0.05, knots = -999)
}
\arguments{
\item{data}{A data frame in the long (interval) format with one line per unit of time.}

\item{Type}{A vector consisting the name of variables representing the start and stop of each
time interval and event indicator. e.g. c("start","stop","event").}

\item{variables}{A vector consisting the name of varialbes that will be adjusted in the model.}

\item{continuous}{A vector of binary indicators representing whether the corresponding variables
are continuous variables (1=Yes, 0=NO)}

\item{TD}{A vector representing whether to force the corresponding variable having a time-dependent
effect. Only takes three values:\cr
1 = force TD effect of the corresponding variable\cr
0 = do not force any effect to the corresponding variable \cr
-1 = force the PH effect of the corresponding variable \cr}

\item{NL}{A vector representing whether to force the corresponding variable having a non-linear
effect. Only takes three values:\cr
1 = force NL effect of the corresponding variable \cr
0 = do not force any effect to the corresponding variable \cr
-1 = force the linear effect of the corresponding variable \cr}

\item{m}{The number of interior knots used in the regression B-spline, default value is 1}

\item{p}{The degree of the regression B-spline, default value is 2}

\item{alpha_back}{significane level used in the backward elemination process}

\item{knots}{Default value is -999}
}
\description{
Build the model through backward selection
}
