% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoxFlex.R
\name{CoxFlex}
\alias{CoxFlex}
\title{Estimate the TD and/or NL effect(s) of continuous covariate(s) in time-to-event analysis}
\usage{
CoxFlex(data, Type, variables, TD, NL, m, p, knots)
}
\arguments{
\item{data}{A data frame in the long (interval) format with one line per unit of time.}

\item{Type}{A vector consisting the name of variables representing the start and stop of each
time interval and event indicator. e.g. c("start","stop","event").}

\item{variables}{A vector consisting the name of varialbes that will be adjusted in the model.}

\item{TD}{A vector of binary indicators representing whether the corresponding variable has a time-dependent
effect or not (1=Yes, 0=No)}

\item{NL}{A vector of binary indicators representing whether the corresponding variable has a non-linear
effect or not (1=Yes, 0=No)}

\item{m}{The number of interior knots used in the regression B-spline}

\item{p}{The degree of the regression B-spline}

\item{knots}{Default value is -999}
}
\value{
Returns a list of the following items:

\item{PLL}{  partial loglikelihood of the final model}
        \item{NP}{  number of parameters estimated in the model}
        \item{NE}{  number of events in the dataset}
        \item{knots_covariates}{knots for splines of NL effects for corresponding covariates}
        \item{knots_time}{knots for splines of the TD effect}
        \item{coefficients_splines_NL}{estimated splines coefficients for NL effects}
        \item{coefficients_splines_TD}{estimated splines coefficients for TD effects}
        \item{variables}{name of the covariates in the order they are adjusted in the model}
        \item{coef}{estimated coefficients for covariates not having TD and NL effects}
        \item{sd}{standard errors of the corresponding estimated coefficients}
        \item{pval}{p-values for corresponding estimated effects}
}
\description{
Estimate the TD and/or NL effect(s) of continuous covariate(s) in time-to-event analysis
}
